Add basic test and rename files to follow standards
