+++ SingedFeathers/Source/SingedFeathers/CoreGame/TileInfo.cpp
@@ -1,7 +1,7 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 #include "SingedFeathers.h"
-#include "Tile_Info.h"
+#include "TileInfo.h"
 
 const tileInfo grassTileInfo = {
     true, //isFlammable
